Update `.gitignore` to modern standard
I noticed this because a build failed in CI because it used the recently
released Nimble version 11.0, which bumped the minimum supported iOS
version from 11.0 to 13.0 breaking compatibility with this project.

The reason CI fetched the latest Nimble version is that it's configured
to use `main` but when I changed the setting, only the project file
changed.

This made me realize the whole `project.xcworkspace` folder, where the
SPM `Package.resolved` is stored, was ignored.